RecipeModel File
This is the file for the RecipeModel which adds in the hardcoded recipes into the home page on the main screen of the app. It displays the name, image, description, ingredients, directions, category, the date that it was published, and the url of the website where the images of the meals were downloaded.