RecipeView File
This is the file for the RecipeView which works by performing actions such as a ScrollView which allows the user to scroll up and down and implementing the images that correspond to each category where it belongs. It also lets the user adjust the height, width, alignment, spacing, and composition of the text.