AddRecipeView File
This is the AddRecipeView file where the user adds a recipe to the app. It performs various actions such as a checkmark which allows the user to add a recipe manually by typing in the Recipe Name, which Category it is, the Description of the recipe, Ingredients it needs, and what the Directions are. They can then hit the checkmark and the new recipe will be displayed in the category section of their choice.