SettingsView File
This is the file of the SettingsView to display the version of the app, which is the newly-developed version, of the recipe app. If it were an actual published app, it would have many more setting options and functions that configure to what the user needs to make the app work properly such as location, storage, account profile, and the option to delete their account from the app.